requirements.txt
openai
python-dotenv
questionary